put models under spid namespace
