--- app/models/spid-rails/settings/slo.rb
@@ -1,15 +1,0 @@
-module Spid
-  module Rails
-
-    class Settings::Slo < Settings
-
-      def to_hash
-        sso_attributes = sp_attributes.merge(idp_attributes)
-        sso_attributes[:sessionindex] = @session_index
-        sso_attributes
-      end
-
-    end 
-
-  end
-end
