+++ app/models/spid/rails/application_record.rb
